clipboard-apis/async-unsanitized-standard-html-formats-write-read.tentative.https.html

<!doctype html>
<meta charset="utf-8">
<title>Async Clipboard unsanitized HTML write -> Async Clipboard unsanitized HTML read test</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/user-activation.js"></script>
<script>
'use strict';

// This function removes extra spaces between tags in html. For example, the
// following html: "<p> Hello </p>   <body> World </body>" would turn into this
//           html: "<p> Hello </p> <body> World </body>"
// We remove the extra spaces because in html they are considered equivalent,
// but when we are comparing for equality the spaces make a difference.
function reformatHtml(html) {
  const parser = new DOMParser();
  const htmlString =
        parser.parseFromString(html, 'text/html').documentElement.innerHTML;
  const reformattedString = htmlString.replace(/\>\s*\</g, '> <');
  return reformattedString;
}

// Writes a payload with custom content and checks to ensure the correct data
// was written successfully.
promise_test(async t => {
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');

  // Create and write unsanitized version of standard HTML and custom formats.
  const format1 = 'text/html';
  const textInput = '<html><head><style>color:blue</style><head><body><p>Hello</p></body></html>';
  const blobInput1 = new Blob([textInput], {type: 'text/html'});
  const clipboardItemInput = new ClipboardItem(
      {[format1]: blobInput1},
      {unsanitized: [format1]});
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItemInput]);

  // Read unsanitized version of HTML format.
  await waitForUserActivation();
  const clipboardItems = await navigator.clipboard.read();

  assert_equals(clipboardItems.length, 1);
  const clipboardItem = clipboardItems[0];
  assert_true(clipboardItem instanceof ClipboardItem);

  // The standard HTML format should have a sanitized HTML fragment.
  const blobOutput1 = await clipboardItem.getType(format1);
  assert_equals(blobOutput1.type, format1);
  const data1 = await (new Response(blobOutput1)).text();
  const outputHtml = reformatHtml(data1);
  const expectedHtml = '<head> </head> <body> <p style=\"color: rgb(0, 0, 0); font-size: medium; ' +
                  'font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; '+
                  'font-weight: 400; letter-spacing: normal; orphans: 2; text-align: start; ' +
                  'text-indent: 0px; text-transform: none; white-space: normal; widows: 2; ' +
                  'word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration-thickness: initial; '+
                  'text-decoration-style: initial; text-decoration-color: initial;\">Hello</p> </body>';
  const sanitizedExpectedHtml = reformatHtml(expectedHtml);
  assert_equals(outputHtml, sanitizedExpectedHtml);
}, 'Verify write and read unsanitized content to the clipboard given text/html format as input');
</script>
